+++ src/App.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState, useMemo } from "react";
 import { Routes, Navigate, Route } from "react-router-dom";
-import TagPage from "./components/tags/TagPage";
 import LatLongContext from "./context/latitude-longitude-context";
 import PermissionContext from "./context/permission-context";
+import RouteContext from "./context/RouteContext";
 import Info from "./components/info";
 import FrontPage from "./components/FrontPage";
 import RoutePage from "./components/routes/RoutePage";
@@ -12,54 +12,48 @@ import FAQ from "./components/FAQ";
 
 function App() {
   const [geolocationAvailable, setGeolocationAvailable] = useState();
+  const [selectedRoute, setSelectedRoute] = useState(null);
+  const [listOfUnlocked, setListOfUnlocked] = useState([]);
+  const [nextUnlockablePointId, setNextUnlockablePointId] = useState(null);
   const [openStreetMapConsent, setOpenStreetMapConsent] = useState(null);
   const [lat, setLat] = useState(null);
   const [long, setLong] = useState(null);
-  const [heading, setHeading] = useState(null);
-  const [speed, setSpeed] = useState(null);
-  const [hasAllowedGeolocation, setHasAllowedGeolocation] = useState(true);
+  const [userAllowedAccessToGeoLocation, setUserAllowedAccessToGeoLocation] = useState(false);
 
   const contextLatLong = useMemo(
     () => ({
       lat,
       long,
-      heading,
-      speed,
     }),
-    [lat, long, heading, speed],
+    [lat, long],
   );
 
-  const geolocationAvailableContext = useMemo(
-    () => ({
-      geolocationAvailable,
-    }),
-    [geolocationAvailable],
-  );
+  useEffect(() => {
+    const experiencesFromLocalStorage = localStorage.getItem("unlocked-experiences");
+    if (experiencesFromLocalStorage) {
+      // add to existing unlocked steps
+      setListOfUnlocked(JSON.parse(experiencesFromLocalStorage));
+    }
+  }, []);
 
   const updateLocation = () => {
     if (lat === null || long === null) {
       navigator.geolocation.getCurrentPosition((position) => {
-        setHasAllowedGeolocation(true);
         setLat(position.coords.latitude);
         setLong(position.coords.longitude);
-        setHeading(position.coords.heading);
-        setSpeed(position.coords.speed);
       });
     }
     setTimeout(() => {
       navigator.geolocation.getCurrentPosition((position) => {
-        setHasAllowedGeolocation(true);
         setLat(position.coords.latitude);
         setLong(position.coords.longitude);
-        setHeading(position.coords.heading);
-        setSpeed(position.coords.speed);
       });
       updateLocation();
     }, 3000);
   };
 
   const handlePermissions = async () => {
-    if (navigator.permissions && navigator.permissions.query) {
+    if (navigator.permissions?.query) {
       // Not apple
       const permissions = await navigator.permissions.query({
         name: "geolocation",
@@ -68,24 +62,22 @@ function App() {
       setGeolocationAvailable(state);
       permissions.onchange = (event) => {
         if (event.target.state === "granted") {
-          setHasAllowedGeolocation(true);
+          setUserAllowedAccessToGeoLocation(true);
         }
         if (event.target.state === "denied") {
-          setHasAllowedGeolocation(false);
+          setUserAllowedAccessToGeoLocation(false);
         }
-        // todo
-        // eslint-disable-next-line no-shadow
-        const { state } = event.target;
-        setGeolocationAvailable(state);
+        setGeolocationAvailable(event.target.state);
       };
+      console.log(state);
       if (state === "granted") {
         updateLocation();
-        setHasAllowedGeolocation(true);
+        setUserAllowedAccessToGeoLocation(true);
       } else if (state === "prompt") {
         updateLocation();
-        setHasAllowedGeolocation(false);
+        setUserAllowedAccessToGeoLocation(false);
       } else if (state === "denied") {
-        setHasAllowedGeolocation(false);
+        setUserAllowedAccessToGeoLocation(false);
       }
     } else {
       // apple
@@ -100,9 +92,10 @@ function App() {
   useEffect(() => {
     // todo some sort of spinner or some indication that something is happening
     requestPermissions();
-  }, []);
+  }, [userAllowedAccessToGeoLocation]);
 
   useEffect(() => {
+    // Consent for handling data with regards to open street map
     const localStorageConsent = localStorage.getItem("data-consent");
     if (localStorageConsent) {
       setOpenStreetMapConsent(localStorageConsent === "true");
@@ -111,21 +104,29 @@ function App() {
 
   return (
     <>
-      {hasAllowedGeolocation && (
-        <div className="App flex flex-col h-full min-h-screen dark:text-white w-screen p-3 text-zinc-800 bg-zinc-100 dark:bg-zinc-800 overflow-hidden">
-          <LatLongContext.Provider value={contextLatLong}>
-            <PermissionContext.Provider
+      <div className="App flex flex-col h-full pt-32 min-h-screen dark:text-white w-screen pl-3 pr-3 pb-3 text-zinc-800 bg-zinc-100 dark:bg-zinc-800 overflow-hidden">
+        <LatLongContext.Provider value={contextLatLong}>
+          <PermissionContext.Provider
+            value={{
+              userAllowedAccessToGeoLocation,
+              openStreetMapConsent,
+              setOpenStreetMapConsent,
+            }}
+          >
+            <RouteContext.Provider
               value={{
-                geolocationAvailableContext,
-                openStreetMapConsent,
-                setOpenStreetMapConsent,
+                selectedRoute,
+                setSelectedRoute,
+                nextUnlockablePointId,
+                setNextUnlockablePointId,
+                listOfUnlocked,
+                setListOfUnlocked,
               }}
             >
-              <Navbar></Navbar>
+              <Navbar />
               <div className="relative grow overflow-hidden">
                 <Routes>
                   <Route path="/" element={<FrontPage />} />
-                  <Route path="tag/:id" element={<TagPage />} />
                   <Route path="route/:id" element={<RoutePage />} />
                   <Route path="faq" element={<FAQ />} />
                   <Route path="/personal-information-policy" element={<PersonalInformationPolicyPage />} />
@@ -134,11 +135,10 @@ function App() {
                   <Route path="*" element={<Navigate to="/" />} />
                 </Routes>
               </div>
-            </PermissionContext.Provider>
-          </LatLongContext.Provider>
-        </div>
-      )}
-      {!hasAllowedGeolocation && <h1>Geolokation er krævet</h1>}
+            </RouteContext.Provider>
+          </PermissionContext.Provider>
+        </LatLongContext.Provider>
+      </div>
     </>
   );
 }